+++ src/Commands/MakeValidatorCommand.php
@@ -21,11 +21,6 @@ class MakeValidatorCommand extends CreateCheck
     protected $command = 'make:validator';
 
     /**
-     * @var string
-     */
-    protected $filename = 'Validate.stub';
-
-    /**
      * @var boolean
      */
     protected $deprecated = true;