+++ src/Scripts/CreateCheck.php
@@ -3,8 +3,6 @@
 namespace Rougin\Weasley\Scripts;
 
 /**
- * Create Check Command
- *
  * @package Weasley
  *
  * @author Rougin Gutib <rougingutib@gmail.com>
@@ -15,16 +13,6 @@ class CreateCheck extends AbstractMake
      * @var string
      */
     protected $command = 'make:check';
-
-    /**
-     * @var string[]
-     */
-    protected $aliases = array('make:validator');
-
-    /**
-     * @var string
-     */
-    protected $filename = 'Check.stub';
 
     /**
      * @var string
@@ -45,4 +33,23 @@ class CreateCheck extends AbstractMake
      * @var string
      */
     protected $text = 'Creates a new check (validation) class based on Valitron';
+
+    /**
+     * @return \Rougin\Classidy\Classidy
+     */
+    protected function stub()
+    {
+        $class = parent::stub();
+
+        $class->extendsTo('Rougin\Weasley\Check');
+
+        $value = 'array<string, string>';
+
+        $class->addArrayProperty('labels', $value)
+            ->asProtected();
+
+        $class->addArrayProperty('rules', $value);
+
+        return $class->asProtected();
+    }
 }