+++ src/Scripts/CreateHandler.php
@@ -2,9 +2,9 @@
 
 namespace Rougin\Weasley\Scripts;
 
+use Rougin\Classidy\Method;
+
 /**
- * Create Handler Command
- *
  * @package Weasley
  *
  * @author Rougin Gutib <rougingutib@gmail.com>
@@ -15,16 +15,6 @@ class CreateHandler extends AbstractMake
      * @var string
      */
     protected $command = 'make:handler';
-
-    /**
-     * @var string[]
-     */
-    protected $aliases = array('make:middleware');
-
-    /**
-     * @var string
-     */
-    protected $filename = 'Handler.stub';
 
     /**
      * @var string
@@ -45,4 +35,51 @@ class CreateHandler extends AbstractMake
      * @var string
      */
     protected $text = 'Creates a new Slytherin Middleware class';
+
+    /**
+     * @return \Rougin\Classidy\Classidy
+     */
+    protected function stub()
+    {
+        $class = parent::stub();
+
+        // Set the interface to the class ------------------------
+        $name = 'Rougin\Slytherin\Middleware\MiddlewareInterface';
+
+        $class->addInterface($name);
+        // -------------------------------------------------------
+
+        $comment = 'Process an incoming server request and return a response, optionally';
+
+        $comment .= "\n" . 'delegating to the next middleware component to create the response.';
+
+        $method = new Method('process');
+
+        $method->setComment($comment);
+
+        // Set the "$request" argument to the class ------
+        $name = 'Psr\Http\Message\ServerRequestInterface';
+
+        $method->addClassArgument('request', $name);
+        // -----------------------------------------------
+
+        // Set the handler interface to the class -------------
+        $name = 'Rougin\Slytherin\Middleware\HandlerInterface';
+
+        $method->addClassArgument('handler', $name);
+        // ----------------------------------------------------
+
+        // Set the interface to return of the class ---
+        $name = '\Psr\Http\Message\ResponseInterface';
+
+        $method->setReturn($name);
+        // --------------------------------------------
+
+        $method->setCodeLine(function ()
+        {
+            return array('//', '', 'return $handler->handle($request);');
+        });
+
+        return $class->addMethod($method);
+    }
 }