+++ src/Scripts/CreatePackage.php
@@ -2,9 +2,9 @@
 
 namespace Rougin\Weasley\Scripts;
 
+use Rougin\Classidy\Method;
+
 /**
- * Create Package Command
- *
  * @package Weasley
  *
  * @author Rougin Gutib <rougingutib@gmail.com>
@@ -15,16 +15,6 @@ class CreatePackage extends AbstractMake
      * @var string
      */
     protected $command = 'make:package';
-
-    /**
-     * @var string[]
-     */
-    protected $aliases = array('make:integration');
-
-    /**
-     * @var string
-     */
-    protected $filename = 'Package.stub';
 
     /**
      * @var string
@@ -45,4 +35,47 @@ class CreatePackage extends AbstractMake
      * @var string
      */
     protected $text = 'Creates a new Slytherin Integration class';
+
+    /**
+     * @return \Rougin\Classidy\Classidy
+     */
+    protected function stub()
+    {
+        $class = parent::stub();
+
+        // Set the integration interface of the class --------------
+        $name = 'Rougin\Slytherin\Integration\IntegrationInterface';
+
+        $class->addInterface($name);
+        // ---------------------------------------------------------
+
+        $method = new Method('define');
+
+        $method->setComment('Defines the specified package.');
+
+        // Set the container interface for the argument --------
+        $name = 'Rougin\Slytherin\Container\ContainerInterface';
+
+        $method->addClassArgument('container', $name);
+        // -----------------------------------------------------
+
+        // Set the configuration for the argument -----------
+        $name = 'Rougin\Slytherin\Integration\Configuration';
+
+        $method->addClassArgument('config', $name);
+        // --------------------------------------------------
+
+        // Set the interface to return of the class -------------
+        $name = '\Rougin\Slytherin\Container\ContainerInterface';
+
+        $method->setReturn($name);
+        // ------------------------------------------------------
+
+        $method->setCodeLine(function ()
+        {
+            return array('//', '', 'return $container;');
+        });
+
+        return $class->addMethod($method);
+    }
 }