+++ tests/Commands/MakeIntegrationCommandTest.php
@@ -2,49 +2,76 @@
 
 namespace Rougin\Weasley\Commands;
 
+use Rougin\Weasley\Console;
+use Rougin\Weasley\Testcase;
 use Symfony\Component\Console\Tester\CommandTester;
 
-class MakeIntegrationCommandTest extends \Rougin\Weasley\Testcase
+/**
+ * @package Weasley
+ *
+ * @author Rougin Gutib <rougingutib@gmail.com>
+ */
+class MakeIntegrationCommandTest extends Testcase
 {
     /**
-     * @var \Rougin\Weasley\Console
+     * @var \Symfony\Component\Console\Application
      */
-    protected $console;
+    protected $app;
 
     /**
-     * Sets up the console application.
-     *
+     * @var string
+     */
+    protected $name = 'make:integration';
+
+    /**
+     * @return void
+     */
+    public function test_passed_if_created()
+    {
+        // Run the command from the console app ---
+        $command = $this->app->find($this->name);
+
+        $command = new CommandTester($command);
+
+        $data = array('name' => 'TestIntegration');
+
+        $command->execute($data);
+        // ----------------------------------------
+
+        // Prepare the expected file after generation ---
+        $path = '/src/Integrations/TestIntegration.php';
+
+        $expect = $file = $this->getRoot() . $path;
+
+        /** @var string */
+        $expect = file_get_contents($expect);
+
+        $expect = str_replace("\r\n", "\n", $expect);
+        // ----------------------------------------------
+
+        // Find the actual generated file ---------------------
+        $path = '/tests/Fixture/Templates/TestIntegration.php';
+
+        $actual = $this->getRoot() . $path;
+
+        /** @var string */
+        $actual = file_get_contents($actual);
+
+        $actual = str_replace("\r\n", "\n", $actual);
+        // ----------------------------------------------------
+
+        $this->assertEquals($expect, $actual);
+
+        unlink($file);
+    }
+
+    /**
      * @return void
      */
     protected function doSetUp()
     {
-        $this->console = new \Rougin\Weasley\Console;
-    }
+        $app = new Console;
 
-    /**
-     * Tests MakeIntegrationCommand::execute.
-     *
-     * @return void
-     */
-    public function testExecuteMethod()
-    {
-        $command = new CommandTester($this->console->find('make:integration'));
-
-        $command->execute(array('name' => 'TestIntegration'));
-
-        $expected = __DIR__ . '/../../src/Integrations/TestIntegration.php';
-        $filename = $expected;
-        /** @var string */
-        $expected = file_get_contents($expected);
-        $expected = str_replace("\r\n", "\n", $expected);
-
-        $original = __DIR__ . '/../Fixture/Templates/TestIntegration.php';
-        /** @var string */
-        $original = file_get_contents($original);
-        $original = str_replace("\r\n", "\n", $original);
-
-        $this->assertEquals($expected, $original);
-
-        unlink($filename);
+        $this->app = $app->make();
     }
 }