+++ tests/Testcase.php
@@ -13,17 +13,31 @@ use LegacyPHPUnit\TestCase as Legacy;
  */
 class Testcase extends Legacy
 {
-    /** @phpstan-ignore-next-line */
-    public function setExpectedException($exception)
+    /**
+     * @param class-string $exception
+     *
+     * @return void
+     */
+    protected function doExpectException($exception)
     {
+        /** @phpstan-ignore-next-line */
         if (method_exists($this, 'expectException'))
         {
+            /** @phpstan-ignore-next-line */
             $this->expectException($exception);
 
             return;
         }
 
         /** @phpstan-ignore-next-line */
-        parent::setExpectedException($exception);
+        $this->setExpectedException($exception);
+    }
+
+    /**
+     * @return string
+     */
+    protected function getRoot()
+    {
+        return __DIR__ . '/../';
     }
 }